Re-Organize + Store Props on Window Manager
