+++ rhubarb_lipsync/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "3.0.2"
+from rhubarb_lipsync import ops, props, ui, prefs
 
 bl_info = {
     "name": "Rhubarb Lipsync-DEV",
@@ -15,26 +15,15 @@ bl_info = {
 }
 
 
-if "bpy" in locals():
-    import importlib
-
-    if "op_blender_rhubarb" in locals():
-        importlib.reload(op_blender_rhubarb)
-        importlib.reload(pnl_blender_rhubarb)
-        importlib.reload(prefs_blender_rhubarb)
-else:
-    from . import op_blender_rhubarb, pnl_blender_rhubarb, prefs_blender_rhubarb
-
-import bpy
-
-
 def register():
-    op_blender_rhubarb.register()
-    pnl_blender_rhubarb.register()
-    prefs_blender_rhubarb.register()
+    ops.register()
+    ui.register()
+    props.register()
+    prefs.register()
 
 
 def unregister():
-    op_blender_rhubarb.unregister()
-    pnl_blender_rhubarb.unregister()
-    prefs_blender_rhubarb.unregister()
+    ops.unregister()
+    ui.unregister()
+    props.unregister()
+    prefs.unregister()
