+++ rhubarb_lipsync/prefs.py
@@ -11,11 +11,8 @@ class RhubarbAddonPreferences(AddonPreferences):
     executable_path: StringProperty(
         name="Rhubarb lipsync executable",
         subtype="FILE_PATH",
-        default=bpy.utils.user_resource("SCRIPTS")
-        + "/addons/blender-rhubarb-lipsync/bin/rhubarb"
-        + (".exe" if system() == "Windows" else ""),
-    )
-
+        default="C:\\Users\\nalbe\AppData\\Roaming\\Blender Foundation\\Blender\\3.3\scripts\\addons\\blender-rhubarb-lipsync\\rhubarb.exe",  # TODO REPLACE
+    )  # TODO default path doesn't work
     recognizer: EnumProperty(
         name="Recognizer",
         items=[
